// extension/as/assembly/k8s/io/apimachinery/pkg/apis/meta/v1/GroupVersionResource.ts
// Code generated by protoc-gen-as. DO NOT EDIT.
// Versions:
//   protoc-gen-as v1.3.0
//   protoc        v3.20.1

import { Writer, Reader } from "as-proto/assembly";

export class GroupVersionResource {
  static encode(message: GroupVersionResource, writer: Writer): void {
    writer.uint32(10);
    writer.string(message.group);

    writer.uint32(18);
    writer.string(message.version);

    writer.uint32(26);
    writer.string(message.resource);
  }

  static decode(reader: Reader, length: i32): GroupVersionResource {
    const end: usize = length < 0 ? reader.end : reader.ptr + length;
    const message = new GroupVersionResource();

    while (reader.ptr < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.group = reader.string();
          break;

        case 2:
          message.version = reader.string();
          break;

        case 3:
          message.resource = reader.string();
          break;

        default:
          reader.skipType(tag & 7);
          break;
      }
    }

    return message;
  }

  group: string;
  version: string;
  resource: string;

  constructor(group: string = "", version: string = "", resource: string = "") {
    this.group = group;
    this.version = version;
    this.resource = resource;
  }
}